rm .Objects method from Get query
