+++ v4/test/graphql/graphql_helper_test.go
@@ -42,7 +42,6 @@ func GetOnePizza(t *testing.T, c *weaviate.Client) *Pizza {
 
 	resp, err := c.GraphQL().
 		Get().
-		Objects().
 		WithClassName("Pizza").
 		WithFields(fields).
 		Do(context.Background())
